AI-Chatbot/requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
torch
nltk
numpy
wikipedia